IKE-ONU_23CG034083_EMOTIONAL_DETECTION_WEB_APP/requirements.txt
Flask==3.0.3
opencv-python==4.10.0.84
numpy==1.26.4
Pillow==10.3.0
scikit-learn==1.5.2
gunicorn==23.0.0

